+++ source/DependencyInjectionExtensions.Tests/AddImplementation/AddServiceAndImplementationTests.cs
@@ -1,8 +1,9 @@
 ﻿using System.Linq;
+using DependencyInjectionExtensions.AddImplementation;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace DependencyInjectionExtensions.Tests
+namespace DependencyInjectionExtensions.Tests.AddImplementation
 {
     public class AddServiceAndImplementationTests
     {
