+++ source/DependencyInjectionExtensions.Tests/AddImplementation/AddWithoutImplementationServiceCollectionExtensionsTests.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using DependencyInjectionExtensions.AddImplementation;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace DependencyInjectionExtensions.Tests
+namespace DependencyInjectionExtensions.Tests.AddImplementation
 {
     public class AddWithoutImplementationServiceCollectionExtensionsTests
     {
