+++ source/DependencyInjectionExtensions/AddImplementation/SimpleServiceCollection.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace DependencyInjectionExtensions
+namespace DependencyInjectionExtensions.AddImplementation
 {
     public class SimpleServiceCollection : List<ServiceDescriptor>, IServiceCollection
     {
