--- source/DependencyInjectionExtensions/IDecoratorFactory.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace DependencyInjectionExtensions
-{
-    public interface IDecoratorFactory
-    {
-        /// <summary>
-        /// Use this method to create decorator at register time
-        /// </summary>
-        /// <param name="toDecorate"></param>
-        /// <param name="decoratingType">Interface to be decorated</param>
-        /// <param name="serviceProvider"></param>
-        /// <returns>Decorated as implementation of <paramref name="decoratingType"/></returns>
-
-        //To pass IServiceProvider is the service locator anti pattern. But we using it only at register time, so it should be ok
-        object CreateDecorated(object toDecorate, Type decoratingType, IServiceProvider serviceProvider);
-
-        bool CanDecorate(Type decoratingType);
-    }
-}
